% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy.birthday.R
\name{happy.birthday}
\alias{happy.birthday}
\title{'Happy Birthday' tune in R}
\usage{
happy.birthday(tune = "normal")
}
\arguments{
\item{tune}{character string. Allows for alternative arrangements.
Can be set to "normal" for the usual tune,
"exciting" for the use of a triplet,
and "very exciting" for the use of a triplet and cadenza.}
}
\description{
Play "Happy Birthday" using R.
}
\details{
The function \emph{\code{happy.birthday}} plays "Happy Birthday" in R.
}
\examples{
\dontrun{
happy.birthday("normal")
}
}
\references{
H/t to Jochen Schäfer on CorrelAid's \url{https://correlaid.org/blog/music-with-r/}!
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{\url{https://chrisdworschak.com/}}
}
