% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merry.christmas.R
\name{merry.christmas}
\alias{merry.christmas}
\title{'Merry Christmas' tune in R}
\usage{
merry.christmas(tune = "normal")
}
\arguments{
\item{tune}{character string. Allows for alternative arrangements.
Can be set to "normal" for the usual tune, and
"exciting" for some alternate notes and rythms.}
}
\description{
Play "Merry Christmas" using R.
}
\details{
The function \emph{\code{merry.christmas}} plays "Merry Christmas" in R.
}
\examples{
\dontrun{
merry.christmas("normal")
}
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{\url{https://chrisdworschak.com/}}
}
